refactor: move SimpleUserCreationForm from models.py to forms.py and improve error messages
