+++ users/templates/users/register.html
@@ -110,7 +110,6 @@
                 placeholder="密碼須包含英文及數字"
                 class="w-full px-3 py-2 pr-10 leading-tight text-gray-700 border rounded shadow appearance-none focus:outline-none focus:shadow-outline"
                 required
-                minlength="8"
               />
               <span
                 onclick="togglePassword('password1')"
@@ -175,7 +174,6 @@
                 placeholder="請再次輸入密碼"
                 class="w-full px-3 py-2 pr-10 leading-tight text-gray-700 border rounded shadow appearance-none focus:outline-none focus:shadow-outline"
                 required
-                minlength="8"
               />
               <span
                 onclick="togglePassword('password2')"
